[android] Moving Firebase MESSAGING_EVENT action into plugin
