+++ push-notifications/android/src/main/AndroidManifest.xml
@@ -1,3 +1,11 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.capacitorjs.plugins.pushnotifications">
+
+    <application>
+        <service android:name="com.capacitorjs.plugins.pushnotifications.MessagingService" android:exported="false">
+            <intent-filter>
+                <action android:name="com.google.firebase.MESSAGING_EVENT" />
+            </intent-filter>
+        </service>
+    </application>
 </manifest>
