+++ push-notifications/android/src/main/java/com/capacitorjs/plugins/pushnotifications/MessagingService.java
@@ -1,0 +1,20 @@
+package com.capacitorjs.plugins.pushnotifications;
+
+import androidx.annotation.NonNull;
+
+import com.google.firebase.messaging.FirebaseMessagingService;
+import com.google.firebase.messaging.RemoteMessage;
+
+public class MessagingService extends FirebaseMessagingService {
+    @Override
+    public void onMessageReceived(@NonNull RemoteMessage remoteMessage) {
+        super.onMessageReceived(remoteMessage);
+        PushNotificationsPlugin.sendRemoteMessage(remoteMessage);
+    }
+
+    @Override
+    public void onNewToken(@NonNull String s) {
+        super.onNewToken(s);
+        PushNotificationsPlugin.onNewToken(s);
+    }
+}
