+++ push-notifications/android/src/main/java/com/capacitorjs/plugins/pushnotifications/PushNotificationsPlugin.java
@@ -8,7 +8,6 @@ import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.service.notification.StatusBarNotification;
-import androidx.annotation.NonNull;
 import com.getcapacitor.Bridge;
 import com.getcapacitor.JSArray;
 import com.getcapacitor.JSObject;
@@ -23,7 +22,6 @@ import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.iid.FirebaseInstanceId;
 import com.google.firebase.iid.InstanceIdResult;
 import com.google.firebase.messaging.FirebaseMessaging;
-import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
 import java.util.ArrayList;
 import java.util.List;
@@ -41,24 +39,6 @@ public class PushNotificationsPlugin extends Plugin {
 
     private static final String EVENT_TOKEN_CHANGE = "registration";
     private static final String EVENT_TOKEN_ERROR = "registrationError";
-
-    public static class MessagingService extends FirebaseMessagingService {
-
-        @Override
-        public void onMessageReceived(@NonNull RemoteMessage remoteMessage) {
-            super.onMessageReceived(remoteMessage);
-            PushNotificationsPlugin pushPlugin = PushNotificationsPlugin.getPushNotificationsInstance();
-            if (pushPlugin != null) {
-                pushPlugin.fireNotification(remoteMessage);
-            }
-        }
-
-        @Override
-        public void onNewToken(@NonNull String s) {
-            super.onNewToken(s);
-            PushNotificationsPlugin.onNewToken(s);
-        }
-    }
 
     public void load() {
         notificationManager = (NotificationManager) getActivity().getSystemService(Context.NOTIFICATION_SERVICE);
